thermal-daemon: Adding the thermal-daemon config files
thermal-conf.xml and thermal-cpu-cdev-order.xml files are the
config files thermal-daemon expects to be defined.

thermal-conf.xml is the file that defines the thermal throttling
policies for a platform. Passive and critical policies are defined
for celadon.

Passive policy controls the cpu power limit when the trip is breached.
Critical policy syncs the file system and shutdown the system upon trip
point breach.

Change-Id: I2bece957a2119011352fdd1cd28ab3dd63339e18
Tracked-On: OAM-67241
Signed-off-by: ysiyer <yegnesh.s.iyer@intel.com>