Kernel : output kernel log to usb serial port
Only support the FTDI type usb-to-serial device.When the device is
inserted into the system, the device is matched by generic_probe of
generic_driver,then the usbserial module waits for the device to send
the data. After receiving the data, it pushes it into the buffer of the
upper TTY device, and the TTY device will send the data to the user, the
printk module can read the data through console module and display it in
minicom terminal.

Specify which console port was used at mixins.spec

Tracked-On: OAM-76637
Signed-off-by: Yanhongx.Zhou <yanhongx.zhou@intel.com>