mixin updates
graphics/media: Handle case when fb is set to i915drmfb
When framebuffer is set to i915drmfb, graphics and media
properties are not set resulting in device not booting and
audio/video sync issues.
Fix the issue by setting hwcomposer, gralloc and codec
properties for frame buffer i915drmfb case.
Dynamic assignment of USB params for camera
Patch identifies USB bus and port number on which
USB camera connects and provides as param to qemu

Tracked-On: OAM-90359
Tracked-On: OAM-90532
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/542
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/562
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Yang, Dong <dong.yang@intel.com>
Signed-off-by: kbillore <kaushal.billore@intel.com>