mixin updates
UVC-gadget: Enable UVC-gadget for CIC target
This patch adds UVC-gadget app in Makefile and
configFS changes in init.rc to add support for CIC
to act as a camera streaming device. Also, the changes
for ACM/NCM which were there for cic_dev target were
not ported to mixins. This patch adds those changes too.

Tracked-On: OAM-89993
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/638
Signed-off-by: saranya <saranya.gopal@intel.com>