mixin updates
Enable device suspend through mixins
This patch allows device suspend
[PM] Handle multiple suspend scenarios in guest
This patch is to handle the scenario if the guest sends multiple suspend events to Host.
When the host is already in suspend mode, and receives another suspend
command, we are just waking the guest.
Every time the guest wakes up, suspend_cnt values is toggled.
This ensures that multiple suspend commands are not sent to the Host.

Tracked-On:
Tracked-On: OAM-91610
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/783
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia/pull/912
Signed-off-by: Shwetha B <shwetha.b@intel.com>