mixin updates
Integrate the openCL relevant so files for caas target.
The openCL relevant so files need to be added into mixin group
of graphic/auto/product.mk.
Increase the super_partition_size for opencl so files integration.
Existing partition size is too small to include opencl so files.
This caused the pre-integration build failure. So we extend the
size accordingly.

Tracked-On: OAM-92070
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/838
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/845
Signed-off-by: Wan Shuang <shuang.wan@intel.com>