mixin updates
Disable FBE if multi-user feature is enabled
Currently multi user feature cannot co-exist with FBE feature,
so previously a patch is committed to disable FBE if multi user
is enabled. However if the '-j' arg is not applied to setup-aic,
the value of MULTIPLE_USER is empty, not false. So FBE is still not
enabled when the multi user feature is not enabled.
Need to modify the judge logic to enable FBE when multi user is
not enabled under the secure mode.

Tracked-On: OAM-92614
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/858
Signed-off-by: ji, zhenlong z <zhenlong.z.ji@intel.com>