mixin updates
Switch to omx software codec
Video format is not supported by c2 on VirtIO GPU,
change back to omx software codec.
Enable sw_omx_video on caas_cfc
This is to help enable sw_omx_video on caas_cfc
and disable it on caas.
Enable uio and kvmfr
This is to help enable uio & kvmmfr config

Tracked-On: OAM-98101
Tracked-On: OAM-98135
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/1188
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/1191
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia/pull/1234
Signed-off-by: Chen Tianmi <tianmi.chen@intel.com>
Signed-off-by: Ren Chenglei <chenglei.ren@intel.com>