mixin updates
Upgrade boot/vendorboot image version to v4
Bootconfig also needs to be enabled for boot/vendorboot v4.
If bootconfig is enabled, all kernel command line parameters
prefixed with 'androidboot' need to be moved to bootconfig.
Add an option for vendor-boot group
Add an option for vendor-boot group to control the bootconfig
feature.

Tracked-On: OAM-100125
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/1268
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia/pull/1314
Signed-off-by: ji, zhenlong z <zhenlong.z.ji@intel.com>