mixin updates
Set the hash algorithm to sha256 for hashtree footer
The default hash algorithm of hashtree footer is sha1,
but this algorithm is prohibited in vts.
Incorporate bootctrl HAL module into recovery image
Currently OTA cannot pass in recovery mode since bootctrl
HAL module is not included in the recovery image.

Tracked-On: OAM-102639
Tracked-On: OAM-102701
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/1413
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/1415
Signed-off-by: jizhenlo <zhenlong.z.ji@intel.com>