mixin updates
Fix ivi software cannot boot up on ADL platform
As GNSS HAL is not yet implemented, GNSS service is not created.
However, the power management function for gnss is enabled currently.
When power state changes, the suspend callback will
be called on gnss service resulting in null pointer dereference.
As GNSS HAL is not implemented, power management for gnss is disabled.

Tracked-On: OAM-103707
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/1453
Signed-off-by: jizhenlo <zhenlong.z.ji@intel.com>