mixin updates
Remove the dell headset entry
Removing the additional headset option(specific to ALC256).
As this was set as a generic option which applies
to all the codecs, this may affect the functionality
of headsets on other codecs in some of platforms.
We should have specific quirks in particular codec files.

Tracked-On: OAM-104331
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/1504
Signed-off-by: Deepa K G <g.k.deepa@intel.com>