mixin updates
Enable kernel configs for Android platform hardening
Kernel drivers needed for Android platform hardening not enabled.
For Android platform hardening, clang and initialization of stack
with zero value is must but the drivers for that is not enabled.
Enable CONFIG_CFI_CLANG and CONFIG_INIT_STACK_ALL_ZERO for
android platform hardening.

Tracked-On: OAM-104319
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/1509
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>