mixin updates
Fix android not booting in sriov mode due to GuC and HuC both enabled
when enable_guc is not overrided, GuC submission and HuC authentication
is enabled which is causing Android Guest boot failure in VF mode.
Fix the issue by overriding the enable_guc to 1 only for SRIOV support
enabled kernel.

Tracked-On: OAM-105298
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/1563
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>