mixin updates
add a new variable of TEE to store tee mixin configuration
this variable 'TEE' will be used to decide packing of optee prebuilt
binary into bootloader. In future it can be used for other decision
making too.
Enable serial console in recovery mode
This can help with debugging issues in userdebug build, in
user build the serial is disabled.
Specify bootloader a slot during flashing
slot a should be used during the flashing.

Tracked-On: OAM-112711
Tracked-On: OAM-112796
Tracked-On: OAM-112901
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/1791
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/1798
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/1799
Signed-off-by: Yan, Shaopu <shaopu.yan@intel.com>
Signed-off-by: jizhenlo <zhenlong.z.ji@intel.com>